+++ CHANGELOG.md
@@ -4,9 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [Unreleased]
+
+### Added
+- Reverse rule default back to the equation of the original rule in case
+  `NotImplementedError`
+
 ### Fixed
 - Removed a debug print
+
 
 ## [3.0.0] - 2021-01-04
 ### Added